+++ include/roq/fbs/api.h
@@ -2673,7 +2673,8 @@ struct MarketByOrderUpdate FLATBUFFERS_FINAL_CLASS : private flatbuffers::Table 
     VT_EXCHANGE_TIME_UTC = 16,
     VT_EXCHANGE_SEQUENCE = 18,
     VT_PRICE_DECIMALS = 20,
-    VT_QUANTITY_DECIMALS = 22
+    VT_QUANTITY_DECIMALS = 22,
+    VT_CHECKSUM = 24
   };
   uint16_t stream_id() const { return GetField<uint16_t>(VT_STREAM_ID, 0); }
   const flatbuffers::String *exchange() const { return GetPointer<const flatbuffers::String *>(VT_EXCHANGE); }
@@ -2695,6 +2696,7 @@ struct MarketByOrderUpdate FLATBUFFERS_FINAL_CLASS : private flatbuffers::Table 
   roq::fbs::Decimals quantity_decimals() const {
     return static_cast<roq::fbs::Decimals>(GetField<uint8_t>(VT_QUANTITY_DECIMALS, 0));
   }
+  uint32_t checksum() const { return GetField<uint32_t>(VT_CHECKSUM, 0); }
   bool Verify(flatbuffers::Verifier &verifier) const {
     return VerifyTableStart(verifier) && VerifyField<uint16_t>(verifier, VT_STREAM_ID) &&
            VerifyOffset(verifier, VT_EXCHANGE) && verifier.VerifyString(exchange()) &&
@@ -2703,7 +2705,8 @@ struct MarketByOrderUpdate FLATBUFFERS_FINAL_CLASS : private flatbuffers::Table 
            verifier.VerifyVector(asks()) && verifier.VerifyVectorOfTables(asks()) &&
            VerifyField<uint8_t>(verifier, VT_UPDATE_TYPE) && VerifyField<int64_t>(verifier, VT_EXCHANGE_TIME_UTC) &&
            VerifyField<int64_t>(verifier, VT_EXCHANGE_SEQUENCE) && VerifyField<uint8_t>(verifier, VT_PRICE_DECIMALS) &&
-           VerifyField<uint8_t>(verifier, VT_QUANTITY_DECIMALS) && verifier.EndTable();
+           VerifyField<uint8_t>(verifier, VT_QUANTITY_DECIMALS) && VerifyField<uint32_t>(verifier, VT_CHECKSUM) &&
+           verifier.EndTable();
   }
 };
 
@@ -2739,6 +2742,7 @@ struct MarketByOrderUpdateBuilder {
   void add_quantity_decimals(roq::fbs::Decimals quantity_decimals) {
     fbb_.AddElement<uint8_t>(MarketByOrderUpdate::VT_QUANTITY_DECIMALS, static_cast<uint8_t>(quantity_decimals), 0);
   }
+  void add_checksum(uint32_t checksum) { fbb_.AddElement<uint32_t>(MarketByOrderUpdate::VT_CHECKSUM, checksum, 0); }
   explicit MarketByOrderUpdateBuilder(flatbuffers::FlatBufferBuilder &_fbb) : fbb_(_fbb) { start_ = fbb_.StartTable(); }
   flatbuffers::Offset<MarketByOrderUpdate> Finish() {
     const auto end = fbb_.EndTable(start_);
@@ -2758,10 +2762,12 @@ inline flatbuffers::Offset<MarketByOrderUpdate> CreateMarketByOrderUpdate(
     int64_t exchange_time_utc = 0,
     int64_t exchange_sequence = 0,
     roq::fbs::Decimals price_decimals = roq::fbs::Decimals_Undefined,
-    roq::fbs::Decimals quantity_decimals = roq::fbs::Decimals_Undefined) {
+    roq::fbs::Decimals quantity_decimals = roq::fbs::Decimals_Undefined,
+    uint32_t checksum = 0) {
   MarketByOrderUpdateBuilder builder_(_fbb);
   builder_.add_exchange_sequence(exchange_sequence);
   builder_.add_exchange_time_utc(exchange_time_utc);
+  builder_.add_checksum(checksum);
   builder_.add_asks(asks);
   builder_.add_bids(bids);
   builder_.add_symbol(symbol);
@@ -2784,7 +2790,8 @@ inline flatbuffers::Offset<MarketByOrderUpdate> CreateMarketByOrderUpdateDirect(
     int64_t exchange_time_utc = 0,
     int64_t exchange_sequence = 0,
     roq::fbs::Decimals price_decimals = roq::fbs::Decimals_Undefined,
-    roq::fbs::Decimals quantity_decimals = roq::fbs::Decimals_Undefined) {
+    roq::fbs::Decimals quantity_decimals = roq::fbs::Decimals_Undefined,
+    uint32_t checksum = 0) {
   auto exchange__ = exchange ? _fbb.CreateString(exchange) : 0;
   auto symbol__ = symbol ? _fbb.CreateString(symbol) : 0;
   auto bids__ = bids ? _fbb.CreateVector<flatbuffers::Offset<roq::fbs::MBOUpdate>>(*bids) : 0;
@@ -2800,7 +2807,8 @@ inline flatbuffers::Offset<MarketByOrderUpdate> CreateMarketByOrderUpdateDirect(
       exchange_time_utc,
       exchange_sequence,
       price_decimals,
-      quantity_decimals);
+      quantity_decimals,
+      checksum);
 }
 
 struct MarketByPriceUpdate FLATBUFFERS_FINAL_CLASS : private flatbuffers::Table {
@@ -2815,7 +2823,8 @@ struct MarketByPriceUpdate FLATBUFFERS_FINAL_CLASS : private flatbuffers::Table 
     VT_EXCHANGE_TIME_UTC = 16,
     VT_EXCHANGE_SEQUENCE = 18,
     VT_PRICE_DECIMALS = 20,
-    VT_QUANTITY_DECIMALS = 22
+    VT_QUANTITY_DECIMALS = 22,
+    VT_CHECKSUM = 24
   };
   uint16_t stream_id() const { return GetField<uint16_t>(VT_STREAM_ID, 0); }
   const flatbuffers::String *exchange() const { return GetPointer<const flatbuffers::String *>(VT_EXCHANGE); }
@@ -2837,6 +2846,7 @@ struct MarketByPriceUpdate FLATBUFFERS_FINAL_CLASS : private flatbuffers::Table 
   roq::fbs::Decimals quantity_decimals() const {
     return static_cast<roq::fbs::Decimals>(GetField<uint8_t>(VT_QUANTITY_DECIMALS, 0));
   }
+  uint32_t checksum() const { return GetField<uint32_t>(VT_CHECKSUM, 0); }
   bool Verify(flatbuffers::Verifier &verifier) const {
     return VerifyTableStart(verifier) && VerifyField<uint16_t>(verifier, VT_STREAM_ID) &&
            VerifyOffset(verifier, VT_EXCHANGE) && verifier.VerifyString(exchange()) &&
@@ -2845,7 +2855,8 @@ struct MarketByPriceUpdate FLATBUFFERS_FINAL_CLASS : private flatbuffers::Table 
            verifier.VerifyVector(asks()) && verifier.VerifyVectorOfTables(asks()) &&
            VerifyField<uint8_t>(verifier, VT_UPDATE_TYPE) && VerifyField<int64_t>(verifier, VT_EXCHANGE_TIME_UTC) &&
            VerifyField<int64_t>(verifier, VT_EXCHANGE_SEQUENCE) && VerifyField<uint8_t>(verifier, VT_PRICE_DECIMALS) &&
-           VerifyField<uint8_t>(verifier, VT_QUANTITY_DECIMALS) && verifier.EndTable();
+           VerifyField<uint8_t>(verifier, VT_QUANTITY_DECIMALS) && VerifyField<uint32_t>(verifier, VT_CHECKSUM) &&
+           verifier.EndTable();
   }
 };
 
@@ -2881,6 +2892,7 @@ struct MarketByPriceUpdateBuilder {
   void add_quantity_decimals(roq::fbs::Decimals quantity_decimals) {
     fbb_.AddElement<uint8_t>(MarketByPriceUpdate::VT_QUANTITY_DECIMALS, static_cast<uint8_t>(quantity_decimals), 0);
   }
+  void add_checksum(uint32_t checksum) { fbb_.AddElement<uint32_t>(MarketByPriceUpdate::VT_CHECKSUM, checksum, 0); }
   explicit MarketByPriceUpdateBuilder(flatbuffers::FlatBufferBuilder &_fbb) : fbb_(_fbb) { start_ = fbb_.StartTable(); }
   flatbuffers::Offset<MarketByPriceUpdate> Finish() {
     const auto end = fbb_.EndTable(start_);
@@ -2900,10 +2912,12 @@ inline flatbuffers::Offset<MarketByPriceUpdate> CreateMarketByPriceUpdate(
     int64_t exchange_time_utc = 0,
     int64_t exchange_sequence = 0,
     roq::fbs::Decimals price_decimals = roq::fbs::Decimals_Undefined,
-    roq::fbs::Decimals quantity_decimals = roq::fbs::Decimals_Undefined) {
+    roq::fbs::Decimals quantity_decimals = roq::fbs::Decimals_Undefined,
+    uint32_t checksum = 0) {
   MarketByPriceUpdateBuilder builder_(_fbb);
   builder_.add_exchange_sequence(exchange_sequence);
   builder_.add_exchange_time_utc(exchange_time_utc);
+  builder_.add_checksum(checksum);
   builder_.add_asks(asks);
   builder_.add_bids(bids);
   builder_.add_symbol(symbol);
@@ -2926,7 +2940,8 @@ inline flatbuffers::Offset<MarketByPriceUpdate> CreateMarketByPriceUpdateDirect(
     int64_t exchange_time_utc = 0,
     int64_t exchange_sequence = 0,
     roq::fbs::Decimals price_decimals = roq::fbs::Decimals_Undefined,
-    roq::fbs::Decimals quantity_decimals = roq::fbs::Decimals_Undefined) {
+    roq::fbs::Decimals quantity_decimals = roq::fbs::Decimals_Undefined,
+    uint32_t checksum = 0) {
   auto exchange__ = exchange ? _fbb.CreateString(exchange) : 0;
   auto symbol__ = symbol ? _fbb.CreateString(symbol) : 0;
   auto bids__ = bids ? _fbb.CreateVector<flatbuffers::Offset<roq::fbs::MBPUpdate>>(*bids) : 0;
@@ -2942,7 +2957,8 @@ inline flatbuffers::Offset<MarketByPriceUpdate> CreateMarketByPriceUpdateDirect(
       exchange_time_utc,
       exchange_sequence,
       price_decimals,
-      quantity_decimals);
+      quantity_decimals,
+      checksum);
 }
 
 struct MarketStatus FLATBUFFERS_FINAL_CLASS : private flatbuffers::Table {