+++ include/roq/fbs/encode.h
@@ -501,7 +501,8 @@ auto encode(B &builder, const roq::MarketByPriceUpdate &value) {
       encode(builder, value.exchange_time_utc),
       value.exchange_sequence,
       encode(builder, value.price_decimals),
-      encode(builder, value.quantity_decimals));
+      encode(builder, value.quantity_decimals),
+      value.checksum);
 }
 
 template <typename B>
@@ -517,7 +518,8 @@ auto encode(B &builder, const roq::MarketByOrderUpdate &value) {
       encode(builder, value.exchange_time_utc),
       value.exchange_sequence,
       encode(builder, value.price_decimals),
-      encode(builder, value.quantity_decimals));
+      encode(builder, value.quantity_decimals),
+      value.checksum);
 }
 
 template <typename B>