+++ include/roq/market_by_price_update.h
@@ -36,6 +36,7 @@ struct ROQ_PUBLIC MarketByPriceUpdate final {
   int64_t exchange_sequence = {};                   //!< Latest sequence number (from exchange)
   Decimals price_decimals = {};                     //!< Decimal digits required to represent prices
   Decimals quantity_decimals = {};                  //!< Decimal digits required to represent quantities
+  uint32_t checksum = {};                           //!< Checksum (internal)
 };
 
 }  // namespace roq
@@ -61,7 +62,8 @@ struct fmt::formatter<roq::MarketByPriceUpdate> {
         R"(exchange_time_utc={}, )"
         R"(exchange_sequence={}, )"
         R"(price_decimals={}, )"
-        R"(quantity_decimals={})"
+        R"(quantity_decimals={}, )"
+        R"(checksum={})"
         R"(}})"sv,
         value.stream_id,
         value.exchange,
@@ -72,7 +74,8 @@ struct fmt::formatter<roq::MarketByPriceUpdate> {
         value.exchange_time_utc,
         value.exchange_sequence,
         value.price_decimals,
-        value.quantity_decimals);
+        value.quantity_decimals,
+        value.checksum);
   }
 };
 template <>